Setting up MongoDB Instance and Creating
