+++ backend/.env.template
@@ -1,1 +1,2 @@
 SESSION_SECRET_KEY=
+URI=
